take-home-test-backend/pyproject.toml
[tool.poetry]
name = "take-home-test-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
greenlet = "1.1.2"
markupsafe = "2.0.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
mypy = "^0.790"
faker = "^4.14.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
